Task 2 - Role-Based Access Control System/backend/requirements.txt
fastapi==0.100.1
uvicorn==0.34.2
sqlalchemy==2.0.41
pydantic==2.11.5
passlib[bcrypt]==1.7.4
python-jose==3.5.0
casbin==1.43.0